requirements.txt
asgiref==3.8.1
Django==5.1.7
class-registry==2.1.2
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
env==0.1.0
filters==1.3.2
inflection==0.5.1
mysqlclient==2.2.7
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
six==1.17.0
sqlparse==0.5.0
uritemplate==4.1.1
